Moved setupPlayer method to the SetupService class and added a check to see if the players ship list is complete.
